# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HireForward — AI-Native Recruitment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /settings:
    get:
      operationId: getCompanySettings
      tags: [settings]
      summary: Get company settings for authenticated company
      responses:
        "200":
          description: Company settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanySettings"
        "401":
          description: Unauthorized
    put:
      operationId: updateCompanySettings
      tags: [settings]
      summary: Update company settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCompanySettingsBody"
      responses:
        "200":
          description: Updated company settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanySettings"
        "401":
          description: Unauthorized

  # ── Job Processes ──────────────────────────────────────────────────────────
  /processes:
    get:
      operationId: listProcesses
      tags: [processes]
      summary: List all job processes for the company
      responses:
        "200":
          description: List of job processes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobProcess"
    post:
      operationId: createProcess
      tags: [processes]
      summary: Create a new job process
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProcessBody"
      responses:
        "201":
          description: Created process
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobProcess"
  /processes/{id}:
    get:
      operationId: getProcess
      tags: [processes]
      summary: Get a job process by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job process
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobProcessDetail"
        "404":
          description: Not found
    patch:
      operationId: updateProcess
      tags: [processes]
      summary: Update a job process
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProcessBody"
      responses:
        "200":
          description: Updated process
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobProcess"
    delete:
      operationId: deleteProcess
      tags: [processes]
      summary: Delete a job process
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /processes/{id}/duplicate:
    post:
      operationId: duplicateProcess
      tags: [processes]
      summary: Duplicate a job process
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Duplicated process
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobProcess"
...
  /processes/{id}/configure:
    post:
      operationId: configureProcessMessage
      tags: [processes]
      summary: Send a message to the configurator agent (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigureMessageBody"
      responses:
        "200":
          description: SSE stream of configurator agent response
          content:
            text/event-stream: {}
  /processes/{id}/finalize-config:
    post:
      operationId: finalizeProcessConfig
      tags: [processes]
      summary: Finalize agent config (system prompt + criteria) and activate process
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Finalized process config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobProcessDetail"

  # ── Candidates ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    JobProcess:
      type: object
      properties:
        id:
          type: string
        companyId:
          type: string
        title:
          type: string
        area:
          type: string
        seniority:
          type: string
        status:
          type: string
          enum: [draft, active, paused, closed]
        interviewType:
          type: string
          enum: [technical, behavioral, hybrid]
        toolsAllowed:
          type: boolean
        responseTimeLimitSeconds:
          type: integer
          nullable: true
        candidatesTotal:
          type: integer
        candidatesEvaluated:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, companyId, title, area, seniority, status, interviewType, toolsAllowed, candidatesTotal, candidatesEvaluated, createdAt]
    JobProcessDetail:
      allOf:
        - $ref: "#/components/schemas/JobProcess"
        - type: object
          properties:
            agentSystemPrompt:
              type: string
            configMessages:
              type: array
              items:
                type: object
                properties:
                  role:
                    type: string
                  content:
                    type: string
                required: [role, content]
            criteria:
              type: array
              items:
                $ref: "#/components/schemas/EvaluationCriterion"
    CreateProcessBody:
      type: object
      properties:
        title:
          type: string
        area:
          type: string
        seniority:
          type: string
        interviewType:
          type: string
          enum: [technical, behavioral, hybrid]
        toolsAllowed:
          type: boolean
      required: [title, area, seniority]
    UpdateProcessBody:
      type: object
      properties:
        title:
          type: string
        area:
          type: string
        seniority:
          type: string
        status:
          type: string
          enum: [draft, active, paused, closed]
        interviewType:
          type: string
          enum: [technical, behavioral, hybrid]
        toolsAllowed:
          type: boolean
        responseTimeLimitSeconds:
          type: integer
          nullable: true
        agentSystemPrompt:
          type: string
    ConfigureMessageBody:
      type: object
      properties:
        content:
          type: string
      required: [content]
    EvaluationCriterion:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        weight:
          type: integer
        descriptors:
          type: object
          additionalProperties: true
      required: [id, name, weight]
...
    CompanySettings:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        logoUrl:
          type: string
          nullable: true
        plan:
          type: string
        status:
          type: string
          enum: [active, inactive, trial]
        website:
          type: string
          nullable: true
        industry:
          type: string
          nullable: true
        companySize:
          type: string
          nullable: true
        timezone:
          type: string
          nullable: true
        interviewLanguage:
          type: string
          nullable: true
        hrContactName:
          type: string
          nullable: true
        hrContactEmail:
          type: string
          nullable: true
        hrContactPhone:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, plan, status, createdAt]
    UpdateCompanySettingsBody:
      type: object
      properties:
        name:
          type: string
        logoUrl:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        industry:
          type: string
          nullable: true
        companySize:
          type: string
          nullable: true
        timezone:
          type: string
          nullable: true
        interviewLanguage:
          type: string
          nullable: true
        hrContactName:
          type: string
          nullable: true
        hrContactEmail:
          type: string
          nullable: true
        hrContactPhone:
          type: string
          nullable: true
...